+++ packages/rev-models/src/index.ts
@@ -3,7 +3,7 @@ export * from './decorators';
 import * as fields from './fields';
 export { fields };
 
-export { IModel, IModelMeta, IModelManager, IValidationContext } from './models/types';
+export { IModel, IModelMeta, IModelManager, IValidationContext, IDefaultsContext } from './models/types';
 export { ModelManager } from './models/manager';
 export { IMethodContext } from './operations/exec';
 export { IModelOperationResult, ModelOperationResult } from './operations/operationresult';